DEFFECT_IMG-WIDTH, LENGTH, PIXELS, 소숫점 3자리로 변경

+++ NotchingGradeInsp/CDefectInfo.h
@@ -9,9 +9,9 @@ public:
 	int nTabNo;				// Cell 번호
 	int nDefNo;				// 결함 번호.
 	int nDefectMode;		// 결함
-	int nSize;				// 
-	int nSizeX;
-	int nSizeY;
+	double nSize;				// 
+	double nSizeX;
+	double nSizeY;
 	int nMaxBright;
 	int nMinBright;
 	int nAvgBright;
